requirements.txt
asgiref==3.4.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.3.2
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-crontab==0.7.1
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-jwt-refresh-token==0.5
djangorestframework-simplejwt==4.4.0
httplib2==0.19.1
idna==2.10
oauth2client==4.1.3
oauthlib==3.1.1
Pillow==8.3.1
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
uWSGI==2.0.19.1
mysqlclient==2.0.3